refactor(station): 지하철역 프러덕션 코드를 각 역할에 맞는 패키지로 이관한다
- presentation
- application
- domain
- infra
