+++ src/main/java/subway/station/application/StationService.java
@@ -1,7 +1,11 @@
-package subway;
+package subway.station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.domain.Station;
+import subway.station.domain.StationRepository;
+import subway.station.presentation.dto.request.StationRequest;
+import subway.station.presentation.dto.response.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
